修改参数积分类型

+++ internal/request/swap_route_request.go
@@ -15,6 +15,6 @@ type SwapRouteRequest struct {
 	IsAntiMev       bool            `form:"is_anti_mev"`
 	Legacy          bool            `form:"legacy"`
 	SwapType        string          `form:"swap_type" binding:"omitempty,oneof=buy sell"`
-	Points          uint64          `form:"points" binding:"omitempty" example:"200000000"`
+	Points          float64         `form:"points" binding:"omitempty" example:"200000000"`
 	PlatformType    string          `form:"platform_type" binding:"required,oneof=pump raydium game g_external g_points" example:"pump"`
 }
